research/pyproject.toml
[tool.poetry]
name = "research"
version = "0.1.0"
description = "Research"


[project]
name = "research"
version = "0.0.1"
dependencies = [
    "copilotkit>=0.1.56a3",
    "crewai==0.118.0",
    "langchain-openai>=0.2.3",
    "langchain-community>=0.3.1",
    "langchain-anthropic>=0.3.17",
    "langchain-google-genai>=2.0.5",
    "langchain>=0.3.26",
    "openai>=1.52.1",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.31.0",
    "requests>=2.32.3",
    "html2text>=2024.2.26",
    "langchain-core>=0.3.25",
    "langgraph-cli[inmem]>=0.3.3",
    "langgraph-checkpoint-sqlite>=2.0.1",
    "aiosqlite>=0.20.0",
    "aiohttp>=3.9.3",
    "fake-useragent>=2.2.0,<3.0.0",
    "readability-lxml>=0.8.4.1,<0.9.0.0",
    "beautifulsoup4>=4.13.4,<5.0.0",
    "lxml>=6.0.0,<7.0.0",
    "pyowm>=3.3.0,<4.0.0",
    "langgraph==0.4.8"
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
copilotkit = "^0.1.56a3"
crewai = "0.118.0"
langchain-openai = "0.2.3"
langchain-community = "^0.3.1"
langchain-anthropic = "0.3.17"
langchain-google-genai = "2.0.5"
langchain = "0.3.26"
openai = "^1.52.1"
tavily-python = "^0.5.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.31.0"
requests = "^2.32.3"
html2text = "^2024.2.26"
langchain-core = "^0.3.25"
langgraph-cli = {extras = ["inmem"], version = "^0.3.3"}
langgraph-checkpoint-sqlite = "^2.0.1"
aiosqlite = "^0.20.0"
aiohttp = "^3.9.3"
langgraph = "0.4.8"

[tool.poetry.scripts]
demo = "research.demo:main"